requirements.txt
Flask==0.10.1
PyYAML==3.11
argparse==1.4.0
redis==2.10.5
redlock==1.2.0
requests==2.9.1
simplejson==3.8.1
uWSGI==2.0.12
